Move the translation to the hook
